simplify the code using macros

+++ src-tauri/src/utils.rs
@@ -3,12 +3,12 @@ use shakmaty::{fen::Fen, Color, Move};
 use crate::constants::*;
 
 pub enum RatingOrMove {
-    Rating(i16),
+    Rating(i32),
     Move(Move),
 }
 
 /// Calculate the weight for either black or white using the given FEN.
-pub fn get_weight_by_fen(fen: Fen, bot_color: Color) -> i16 {
+pub fn get_weight_by_fen(fen: Fen, bot_color: Color) -> i32 {
     let weight_for_white = fen
         .to_string()
         .split_once(' ')
